use interop test files to test `NSID` core class
