+++ tests/interop-test-files/syntax/aturi_syntax_valid.txt
