+++ tests/interop-test-files/syntax/did_syntax_invalid.txt
