+++ tests/interop-test-files/syntax/nsid_syntax_invalid.txt
