+++ tests/interop-test-files/syntax/recordkey_syntax_invalid.txt
