+++ tests/interop-test-files/syntax/recordkey_syntax_valid.txt
